+++ src/types/index.ts
@@ -1,5 +1,0 @@
-export * from './RouteConfig';
-export * from './RouteContext';
-export * from './RouteError';
-export * from './RouteEvent';
-export * from './RouterConfig';